requirements.txt
Babel==1.3
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
WTForms==2.0
argparse==1.2.1
httpagentparser==1.7.3
httplib2==0.9
oauthlib==0.6.3
python-linkedin==4.1
pytz==2014.4
requests==2.3.0
requests-oauthlib==0.4.1
webapp2==2.5.2
wsgiref==0.1.2
